deleted some delete operators in trueHitTester and trackFitChecker that caused seg faults. Changed mctrackfinder to use gRandom. Deleted the TRandom includes in both genfitter modules. Fromatted the includes in GenFitter.cc a bit nicer.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2439 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/modules/mctrackfinder/include/MCTrackFinderModule.h
@@ -13,6 +13,7 @@
 
 #include <framework/core/Module.h>
 
+#include <string>
 
 namespace Belle2 {
 
@@ -83,7 +84,7 @@ namespace Belle2 {
     bool m_neutrals;                                            /**< Boolean to mark if track candidates should also be created for neutral particles.*/
 
 
-    int m_smearing;                                             /**< Smearing of MCMomentum and MCVertex in % */
+    double m_smearing;                                             /**< Smearing of MCMomentum and MCVertex in % */
 
     std::string m_gfTrackCandsColName;                          /**< TrackCandidates collection name */
     std::string m_gfTrackCandToMCParticleColName;               /**< TrackCandidates to MCParticles relation name */
